Fix delete-button Bug
When a task-bar checkbox is selected, a delete button is displayed.
However, clicking any nav button would display task-bars specified by
the nav button but the delete-button would still remain on display.
This fixes the above mentioned bug and now the delete button is removed
whenever a nav button is clicked.